ceres/requirements.txt
docker
colorama
pyyaml